pyproject.toml
[tool.poetry]
name = "generative"
version = "0.0.1"
description = "Placeholder description"
authors = ["Richard"]
readme = "README.rst"
repository = "https://github.com/samedii/generative"
# some classifiers are added automatically, see https://python-poetry.org/docs/pyproject/#classifiers
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Other Environment",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "generative" },
]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.6.0"
numpy = "^1.19.2"
torchvision = "^0.7.0"
opencv-python = "^4.4.0"
pytorch-datastream = "^0.4.0"
pytorch-lantern = "^0.9.3"

[tool.poetry.dev-dependencies]
guildai = "0.7.1"
data-kale = "^0.1.3"
pylint = "^2.6.0"
flake8 = "^3.8.4"
black = "^20.8b1"
pytest = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
